requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.2,<4.0
daphne>=4.1,<5.0
openpyxl>=3.1,<4.0
django-allauth[socialaccount]>=65.0,<66.0
